git-svn-id: https://svn.qdecoder.org/qdecoder/trunk@279 5a34604f-89d6-db11-b6c0-0015c5f490bc

+++ src/qCgiResponse.c
@@ -212,7 +212,7 @@ int qCgiResponseDownload(Q_ENTRY *request, const char *filepath, const char *mim
 	printf("Content-Disposition: %s;filename=\"%s\"\n", disposition, filename);
 	printf("Content-Transfer-Encoding: binary\n");
 	printf("Accept-Ranges: bytes\n");
-	printf("Content-Length: %ju\n", qFileGetSize(filepath));
+	printf("Content-Length: %zu\n", (size_t)qFileGetSize(filepath));
 	printf("Connection: close\n");
 	qCgiResponseSetContentType(request, mime);
